% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdd_function.R
\name{basic_est}
\alias{basic_est}
\title{Subfunction that creates basic MRD estimates}
\usage{
basic_est(
  Y,
  X1,
  X2,
  W = NULL,
  F1.range = NULL,
  F2.range = NULL,
  R1 = NULL,
  R2 = NULL,
  K = -1,
  save_bayes_se = TRUE,
  fuzzy = FALSE,
  reverse = FALSE,
  sp = NULL,
  save_fits = TRUE,
  save_density_est = TRUE,
  pred_dfs = NULL,
  undersmooth,
  save_sp = TRUE
)
}
\arguments{
\item{Y}{outcome variable}

\item{X1}{first running variable}

\item{X2}{second running variable}

\item{W}{whether the observation is treated. defaults to NULL, but must be specified if the fuzzy option is TRUE}

\item{F1.range}{range for the first running variable, defaults to NULL in which case the range is set to 0 to max(X1) with equal intervals}

\item{F2.range}{range for the second running variable, defaults to NULL in which case the range is set to 0 to max(X2) with equal intervals}

\item{R1}{indices for observations in treated region, defaults to NULL}

\item{R2}{indices for observations in untreated region, defaults to NULL}

\item{K}{parameter controlling the penalization for thin plate regression splines, defaults to -1}

\item{save_bayes_se}{indicate whether Bayesian standard errors should be saved}

\item{fuzzy}{specify whether this is a fuzzy RD, defaults to FALSE}

\item{reverse}{specify whether the running variables are reversed (i.e. negative values are treated), defaults to FALSE}

\item{sp}{parameter for thin plate regression spline fitting}

\item{save_fits}{specify whether fitted thin plate regression splines should be saved, defaults to true}

\item{save_density_est}{specify whether kernel density estimates should be saved, defaults to TRUE}

\item{pred_dfs}{specify a dataframe with the values of the running variables that treatment effects should be predicted at, defaults to NULL}

\item{undersmooth}{specify whether thin plate regression splines should be undersmoothed to reduce bias}

\item{save_sp}{specify whether the parameter for thin plate regregression spline fitting should be saved, defaults to TRUE}
}
\description{
basic_est is a subfunction that does the basic MRD estimation
}
\keyword{basic_mrd_est}
