% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot}
\alias{create_plot}
\title{Function to create basic MRD plot}
\usage{
create_plot(
  x,
  est,
  se,
  legend = "MRD",
  reverse = FALSE,
  my.colors = "black",
  my_xlab,
  my_ylab
)
}
\arguments{
\item{x}{range for running variable}

\item{est}{estimates of the treatment effects}

\item{se}{standard error of estimated treatment effects}

\item{legend}{title for legend}

\item{reverse}{reverse the running variable, defaults to FALSE}

\item{my.colors}{specify desired colors for plot, defaults to black}

\item{my_xlab}{title for x-axis}

\item{my_ylab}{title for y-axis}
}
\description{
create_plot is a function that takes in the range for a running variable,
estimates of the treatment effect and standard errors along that range, and other arguments,
and returns a ggplot object, using the default_ggplot subfunction.
This function is mainly used to generate basic plot for the MRD estimates.
}
\keyword{basic_plot}
