% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{default_ggplot}
\alias{default_ggplot}
\title{Subfunction for plotting}
\usage{
default_ggplot(
  df,
  my.colors,
  my_ylim = NULL,
  my_ylab,
  my_xlab,
  plot_ci = TRUE,
  crit_val = NULL
)
}
\arguments{
\item{df}{ggplot dataframe}

\item{my.colors}{specify desired colors for plot}

\item{my_ylim}{optional numeric vector specifying the limits for the y-axis}

\item{my_ylab}{title for y-axis}

\item{my_xlab}{title for x-axis}

\item{plot_ci}{specify whether we want to plot confidence intervals, defaults to TRUE}
}
\description{
default_ggplot is a subfunction that takes in a ggplot dataframe and
a number of arguments and creates a ggplot object.
}
\keyword{plot_subfn}
