% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot}
\alias{create_plot}
\title{Function to create basic MRD plot}
\usage{
create_plot(
  x,
  est,
  se,
  crit_val = NULL,
  legend = "MRD",
  reverse = FALSE,
  my_ylim = NULL,
  my.colors = "black",
  my_xlab,
  my_ylab
)
}
\arguments{
\item{x}{The input should be a vector containing the points at which the user wants to plot the conditional average treatment effect (CATE). To elaborate, the CATE can only be calculated at the boundary where one of the running variables is zero, and the other takes positive values, and this option specifies at which positive values the CATE should be plotted.}

\item{est}{The input should be a vector containing treatment effect estimates, which should be the same length as x.}

\item{se}{The input should be a vector containing point-wise standard errors for the treatment effect estimates, which should be the same length as x.}

\item{legend}{The input be the title for legend.}

\item{reverse}{The input should be a logical value (TRUE or FALSE) indicating whether the running variable is reversed (i.e. negative values are treated). If not specified, the option defaults to FALSE.}

\item{my_ylim}{optional numeric vector specifying the limits for the y-axis}

\item{my.colors}{The input should be the desired colors for plot. If left empty, the default color is black.}

\item{my_xlab}{The input should be the title of the x-axis.}

\item{my_ylab}{The input should be the title of the y-axis.}
}
\description{
create_plot is a function that takes in the range for a running variable,
estimates of the treatment effect and standard errors along that range, and other arguments,
and returns a ggplot object, using the default_ggplot subfunction.
This function is mainly used to generate basic plot for the MRD estimates.
}
\keyword{basic_plot}
