% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdd_function.R
\name{mrdd.fn}
\alias{mrdd.fn}
\title{Main function for MRD Estimation}
\usage{
mrdd.fn(
  Y,
  X1,
  X2,
  W = NULL,
  F1.range = NULL,
  F2.range = NULL,
  R1 = NULL,
  R2 = NULL,
  zscore = qnorm(0.975),
  K = -1,
  bs = FALSE,
  bs_reps = 1000,
  bs_seeds = TRUE,
  save_bayes_se = TRUE,
  save_density_est = TRUE,
  save_fits = TRUE,
  fuzzy = FALSE,
  reverse = FALSE,
  undersmooth = TRUE,
  pred_dfs = NULL
)
}
\arguments{
\item{Y}{outcome variable}

\item{X1}{first running variable}

\item{X2}{second running variable}

\item{W}{whether the observation is treated. defaults to NULL, but must be specified if the fuzzy option is TRUE}

\item{F1.range}{range for the first running variable, defaults to NULL in which case the range is set to 0 to max(X1) with equal intervals}

\item{F2.range}{range for the second running variable, defaults to NULL in which case the range is set to 0 to max(X2) with equal intervals}

\item{R1}{indices for observations in treated region, defaults to NULL}

\item{R2}{indices for observations in untreated region, defaults to NULL}

\item{zscore}{zscore for statistical significance, defaults to qnorm(0.975) i.e. 5 percent significance level}

\item{K}{parameter controlling the penalization for thin plate regression splines, defaults to -1}

\item{bs}{specify whether to use bootstrap for standard errors, defaults to FALSE}

\item{bs_reps}{number of bootstrap iterations, defaults to 1000}

\item{bs_seeds}{specify whether the random seed for the rth bootstrap iteration should be r, defaults to TRUE}

\item{save_bayes_se}{indicate whether Bayesian standard errors should be saved, defaults to TRUE}

\item{save_density_est}{specify whether kernel density estimates should be saved, defaults to TRUE}

\item{save_fits}{specify whether fitted thin plate regression splines should be saved, defaults to true}

\item{fuzzy}{specify whether this is a fuzzy RD, defaults to FALSE}

\item{reverse}{specify whether the running variables are reversed (i.e. negative values are treated), defaults to FALSE}

\item{undersmooth}{specify whether thin plate regression splines should be undersmoothed to reduce bias}

\item{pred_dfs}{specify a dataframe with the values of the running variables that treatment effects should be predicted at, defaults to NULL}
}
\description{
mrdd.fn is the main function for MRD estimation
}
