% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdd_function.R
\name{mrdd.fn}
\alias{mrdd.fn}
\title{Main function for estimation of regression discontinuity designs with two running variables (MRD)}
\usage{
mrdd.fn(
  Y,
  X1,
  X2,
  W = NULL,
  F1.range = NULL,
  F2.range = NULL,
  mrk = FALSE,
  kink = TRUE,
  sharp_mrk = TRUE,
  sharp_mrd = FALSE,
  eps = 10^(-7),
  R1 = NULL,
  R2 = NULL,
  signif_lvl = 0.05,
  K = -1,
  bs = FALSE,
  bs_reps = 1000,
  bs_seeds = TRUE,
  save_density_est = TRUE,
  save_fits = TRUE,
  fuzzy = FALSE,
  reverse = FALSE,
  fn_ci = TRUE,
  reps_fn_ci = 1000,
  bc = TRUE,
  undersmooth = TRUE,
  pred_dfs = NULL,
  bs_print = 100
)
}
\arguments{
\item{Y}{Y is a vector for the dependent variable.}

\item{X1}{X1 is a vector for the first running variable.}

\item{X2}{X2 is a vector for the second running variable.}

\item{W}{W is a vector containing dummy variables for whether the observation is treated. This defaults to NULL, and it is assumed to be a sharp RD where a unit is treated if and only if both running variables exceed their threshold. If it is in fact a fuzzy RD (so the fuzzy option is specified to be TRUE), then W must be specified.}

\item{F1.range}{This is the points at which the user wants to compute the conditional average treatment effect at the boundary where X2=0, and X1 is positive. The default for this option is NULL, in which case the conditional average treatment effect is computed from X1=0 to max(X1), at equal intervals.}

\item{F2.range}{This is the points at which the user wants to compute the conditional average treatment effect at the boundary where X1=0, and X2 is positive. The default for this option is NULL, in which case the conditional average treatment effect is computed from X2=0 to max(X2), at equal intervals.}

\item{mrk}{This indicates whether we want to estimate a multidimensional regression kink design.}

\item{kink}{This indicates whether we want to estimate a derivative discontinuity at the threshold. This option is typically used when we don't have an MRK/MRDK design but still want to estimate a discontinuity in the derivative.}

\item{sharp_mrk}{This indicates whether this is a sharp MRK design. If this is set to TRUE, then analytic standard errors for the MRK estimate are computed ignoring the estimation uncertainty in the denominator.}

\item{sharp_mrd}{This indicates whether this is a sharp MRD design. If this is set to TRUE but the denominator is also estimated, then analytic standard errors for the MRD estimate are computed ignoring the estimation uncertainty in the denominator.}

\item{eps}{This value indicates the step size used for the numerical derivative computation in the MRK estimation.}

\item{R1}{The input for this option should be a vector containing indices for observations in treated region. Typically, this can be left blank, in which case R1 defaults to NULL.}

\item{R2}{The input for this option should be a vector containing indices for observations in untreated region. Typically, this can be left blank, in which case R2 defaults to NULL.}

\item{signif_lvl}{The indicates the statistical significance level for computing the simultaneous confidence band. If left empty, this defaults to the 5 percent significance level.}

\item{K}{This is a parameter controlling the penalization for thin plate regression splines. Typically this is left empty, in which case K defaults to -1.}

\item{bs}{The input should be a logical value indicating whether to use bootstrap for standard errors. If not specified, this option defaults to FALSE.}

\item{bs_reps}{The input should be the number of desired bootstrap iterations. If not specified, this is option defaults to 1000.}

\item{bs_seeds}{The input should be a logical value indicating whether the random seed for the rth bootstrap iteration should be set to r. If not specified, this option defaults to TRUE.}

\item{save_density_est}{The input should be a logical value indicating whether kernel density estimates should be saved. If the option is not specified, it defaults to TRUE.}

\item{save_fits}{The input should be a logical value indicating whether the fitted thin plate regression splines should be saved. If the option is not specified, it defaults to TRUE.}

\item{fuzzy}{The input should be a logical value (TRUE or FALSE) indicating whether this is a fuzzy RD. If this is not specified, the option defaults to FALSE (so the code assumes that this is a sharp RD).}

\item{reverse}{The input should be a logical value (TRUE or FALSE) indicating whether the treatment is reversed, i.e., what is defined as the "treatment group" for purposes of the MRD due to the region of the running variable space that defines it is is actually the control for the empirical application.}

\item{fn_ci}{This indicates whether simultaneous confidence bands should be computed for the CATE estimates along the treatment frontier.}

\item{reps_fn_ci}{This indicates the number of simulation draws to be used for computing the simultaneous confidence bands for the CATE estimates.}

\item{bc}{The input should be a logical value indicating whether the thin plate regression splines should be undersmoothed by using the MSE-optimal penalty parameter from a higher-order spline as the penalty parameter for the splines actually used to compute the treatment effects.}

\item{undersmooth}{The input should be a logical value indicating whether thin plate regression splines should be undersmoothed by dividing the MSE-optimal penalty parameter by two, in order to reduce bias (for inference purposes).}

\item{pred_dfs}{The input should be a dataframe with the values of the running variables that conditional average treatment effects should be calculated at. If left empty, it defaults to NULL.}

\item{bs_print}{This indicates the multiples at which bootstrap iterations should be printed (if the bs option is set to TRUE).}
}
\value{
The mrdd.fn returns a nested list containing the estimation results. Key components of the returned list include:
\itemize{
\item{"tau"}{This sublist includes the MRD estimates. Furthered nested within this sublist may include the subsublists: "reduced", "stage1", and "late". For sharp MRD designs, only "reduced" is populated and contains the MRD estimate, whereas for fuzzy designs, "stage1" contains estimates for the denominator of the Wald ratio, and "late" is the local average treatment effect/Wald Ratio. Further nested within "reduced" (as well as "stage1" and "late") are the subsubsublists "pw" and "agg", which contain either the pointwise CATE estimates, or the CATE estimates aggregated over certain parts of the treatment frontier respectively.}
\item{"tau_se"}{This sublist contains standard errors for the MRD estimates. It contains similar subsublists to the "tau" sublist, which in turn contains standard errors for the CATE at every point along the treatment frontier ("pw"), or for the aggregated treatment effect estimate ("agg"). Nested within "pw" and "delta" may include "bayes" which corresponds to the Bayesian standard errors (for "pw") or "delta" which corresponds to the delta method (for "agg") which incorporates both uncertainty in the thin plate regression spline estimation and the density estimation, and if the "bs" option in the function is specified to be true, then bootstrap standard errors ("bs") are also available.}
\item{"fn_cov"}{This sublist contains critical values for the simultaneous confidence bands (corresponding to the significance level as specified by the "signif_lvl" option in the mrdd.fn function). For example, while the critical value for constructing a 95 percent pointwise confidence interval is 1.96, it is typically larger for a 95 percent simultaneous confidence band, which should contain the entire CATE over the treatment frontier 95 percent of the time over random samples.}
\item{"fits"}{This sublists contains the fitted thin plate regression splines underlying the MRD/MRK estimates.}
. \item{"density_est" and "density_preds"}{These sublists contain density estimation results and the predicted values of the density function over the treatment frontier, which are used for computing the aggregated treatment effect over the treatment frontier.}
}
Many variations of the sublists mentioned above are included in the brief description above are also included in the output. To better understand what some of them mean, the suffix "bc" and "us" refer to quantities associated with estimation that uses bias-corrected estimates (based on higher-order thin plate regression splines), or undersmoothed estimates (by dividing the MSE-optimal penalty parameter by one-half) respectively. The suffix "mrk" refers to multidimensional regression kink estimates (or if the "mrk" option is FALSE and the "kink" option is TRUE, simply a kink estimate), The terms "F1" and "F2" refer to the region of the treatment frontier where the first running variable is equal to zero and the second is positive, and vice versa, respectively.
The terms "reduced", "stage1", and "late" are borrowed from the instrumental variables literature. For sharp MRD designs, the MRD estimate corresponds to the results under "reduced". For fuzzy designs (and MRK designs), "reduced" and "stage1" correspond to the numerator and denominator of the Wald ratio, and the MRD (or MRK) treatment effect is given by the Wald ratio itself, abbreviated as "late" (short for local average treatment effect).
}
\description{
mrdd.fn is the main function for MRD estimation
}
